+++ globook_server/src/main/java/org/gdsc/globook/domain/type/EUserRole.java
@@ -1,4 +1,4 @@
-package org.gdsc.globook.domain;
+package org.gdsc.globook.domain.type;
 
 import lombok.RequiredArgsConstructor;
